requirements.txt
# Core dependencies for Yahoo Finance AI Browser Agent (Python 3.7+ compatible)
selenium==4.0.0
beautifulsoup4==4.10.0
requests==2.27.1
pandas==1.3.5
numpy==1.19.5
lxml==4.6.5

# AI and ML libraries
openai==0.28.1

# Browser automation
webdriver-manager==3.5.4

# Data processing and analysis
yfinance==0.1.87
matplotlib==3.5.3

# Configuration and logging
python-dotenv==0.19.2
loguru==0.6.0
pydantic==1.8.2

# Utilities
tqdm==4.62.3
click==8.0.4
rich==12.6.0
python-dateutil==2.8.2
